django-server/requirements.txt
django==1.10.3
gunicorn==19.6.0
django-solo==1.1.2
sorl-thumbnail==12.3
django-dotenv==1.4.1
coloredlogs==5.0
https://github.com/martinogden/django-jet/archive/master.zip